requirements.txt
annotated-types @ file:///home/conda/feedstock_root/build_artifacts/annotated-types_1733247046149/work
anyio @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_anyio_1754315087/work
black @ file:///home/conda/feedstock_root/build_artifacts/black-recipe_1742502750715/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1754231422783/work/certifi
click @ file:///home/conda/feedstock_root/build_artifacts/click_1747811314515/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
dnspython @ file:///home/conda/feedstock_root/build_artifacts/dnspython_1733256735222/work
email_validator @ file:///home/conda/feedstock_root/build_artifacts/email-validator-meta_1733300719943/work
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1746947292760/work
fastapi @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_fastapi-core_1752617909/work
fastapi-cli @ file:///home/conda/feedstock_root/build_artifacts/fastapi-cli_1751972066250/work
h11 @ file:///home/conda/feedstock_root/build_artifacts/h11_1745526374115/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_httpcore_1745602916/work
httptools @ file:///home/conda/feedstock_root/build_artifacts/httptools_1732707661751/work
httpx @ file:///home/conda/feedstock_root/build_artifacts/httpx_1733663348460/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
iniconfig @ file:///home/conda/feedstock_root/build_artifacts/iniconfig_1733223141826/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
markdown-it-py @ file:///home/conda/feedstock_root/build_artifacts/markdown-it-py_1754951200865/work
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1733219680183/work
mdurl @ file:///home/conda/feedstock_root/build_artifacts/mdurl_1733255585584/work
mypy_extensions @ file:///home/conda/feedstock_root/build_artifacts/mypy_extensions_1745776566517/work
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1755864419765/work/dist/numpy-2.3.2-cp311-cp311-linux_x86_64.whl#sha256=bddec67bdb90db5b31235716728694a99b7016954300b83f763483467cd3d17d
onnxruntime-gpu @ file:///home/conda/feedstock_root/build_artifacts/onnxruntime_1746973265650/work/build-ci/Release/dist/onnxruntime_gpu-1.22.0-cp311-cp311-linux_x86_64.whl#sha256=d26e3b2cde225279432005edd1be04f5d3939434aa6eb9652cbe37dc240aed7b
opencv-python==4.12.0
opencv-python-headless==4.12.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pathspec @ file:///home/conda/feedstock_root/build_artifacts/pathspec_1733233363808/work
pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1751481992986/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1746710438/work
pluggy @ file:///home/conda/feedstock_root/build_artifacts/pluggy_1747339660894/work
pydantic @ file:///home/conda/feedstock_root/build_artifacts/pydantic_1749927112905/work
pydantic-settings @ file:///home/conda/feedstock_root/build_artifacts/pydantic-settings_1750801564484/work
pydantic_core @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pydantic-core_1746625319/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
pytesseract @ file:///home/conda/feedstock_root/build_artifacts/pytesseract_1736249287101/work
pytest @ file:///home/conda/feedstock_root/build_artifacts/pytest_1750239416491/work
python-dotenv @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dotenv_1750789290/work
python-multipart @ file:///home/conda/feedstock_root/build_artifacts/python-multipart_1734420773152/work
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1737454647378/work
rich @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_rich_1753436991/work/dist
rich-toolkit @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_rich-toolkit_1755001296/work
ruff @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ruff_1755823612/work
shellingham @ file:///home/conda/feedstock_root/build_artifacts/shellingham_1733300899265/work
sniffio @ file:///home/conda/feedstock_root/build_artifacts/sniffio_1733244044561/work
starlette @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_starlette_1753374988/work
tomli @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_tomli_1753796677/work
typer==0.16.1
typer-slim==0.16.1
typing-inspection @ file:///home/conda/feedstock_root/build_artifacts/typing-inspection_1747870647094/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1751643513/work
uvicorn @ file:///home/conda/feedstock_root/build_artifacts/uvicorn_1751201600733/work
uvloop @ file:///home/conda/feedstock_root/build_artifacts/uvloop_1730214322103/work
watchfiles @ file:///home/conda/feedstock_root/build_artifacts/watchfiles_1750053833473/work
websockets @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_websockets_1753459933/work/dist
